[REFACTO] : Changes long ids to Long and corrects methods with int id parameters

+++ src/main/java/com/excilys/formation/persistence/jdbc/ComputerDaoJdbc.java
@@ -60,7 +60,7 @@ public class ComputerDaoJdbc implements ComputerDao {
             preparedStatement.executeUpdate();
             ResultSet resultSet = preparedStatement.getGeneratedKeys();
             if (resultSet != null && resultSet.next()) {
-                int computerId = resultSet.getInt(1);
+                long computerId = resultSet.getLong(1);
                 pComputer.setId(computerId);
                 resultSet.close();
                 return pComputer;
